frogger_app/requirements.txt
pyyaml>=6.0
pandas>=2.2
biopython>=1.83
openpyxl>=3.1
tqdm>=4.66